amazon_aplus_mvp/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
playwright==1.55.0
pydantic==2.11.7
python-dotenv==1.0.1
